+++ src/lib.rs
@@ -228,7 +228,7 @@ mod utility_test;
 /// let predicted = Array1::from(vec![1.0, 0.0, 1.0, 1.0, 0.0]);
 /// let actual = Array1::from(vec![1.0, 0.0, 0.0, 1.0, 1.0]);
 ///
-/// let cm = ConfusionMatrix::new(&predicted, &actual);
+/// let cm = ConfusionMatrix::new(&predicted, &actual).unwrap();
 /// println!("{}", cm.summary());
 /// ```
 pub mod metric;