+++ changelog.md
@@ -9,6 +9,7 @@ Please view [SomeB1oody/RustyML](https://github.com/SomeB1oody/RustyML) for more
 ### Changed
 - Change `principal_component_analysis` module to `utility` module, change `principal_component_analysis_test` module to `utility_test` module
 - Keep the algorithm functions in the math module, and move the functions that evaluate the model's performance (such as R-square values) and structures (confusion matrices) to the metric module. Some of them are used in both ways, then keep them in both modules.
+- Change the output of some of the functions in math module and metric module from `T` to `Result<T, crate::ModelError>`
 
 ## [v0.1.1] - 2025-3-31
 ### Added